requirements.txt
asgiref==3.5.2
Django==3.2.6
pytz==2022.4
sqlparse==0.4.3
psycopg2-binary==2.9.1
djangorestframework==3.11.0
requests==2.28.1
celery==5.1.2
redis==3.5.3